make sure that you can only get to the login /register page if you arent logged in,
In the dashboard the navbar should be fixed to the top and under that there should first be a "Recenty Added" slideshow, the slideshow stretches from left to right and cycles through the latest 5 added videos.

Make sure the backend scans a media_file directory for movie and show files

Make sure the media_files has 2 Directories in it called Movies and Shows, Make sure it beautifies the title of the files so its readable.

Make sure it scanns movie titles for their name, subname (optional) , quality, date of release

Make sure it scanns shows on their name, Season number, Episode Number, episode name (optional)

Make sure for both that it automatically implements the SRT file for the file if its found in a Subtitle_files directory